requirements.txt
openai>=1.3.7
python-dotenv>=1.0.0
Pillow>=10.0.0
torch>=2.1.0
transformers>=4.40.0